media-viewer: add render image with annotations operator.
This is a bigger patch that adds a new gpu_ops and gpu_opsdata module. The render_img_with_annotations operators uses the gpu and bgl module to create a render out of the image editor that contains the annotations that were drawn over it. This was achieved with a lot of help from Jeroen Bakker. Known limitations are:
-Multilayer EXRs: Operator always writes out the first imgae layer, even if user selects another one. (I don't know why that is yet)
-The colorspace of rendered EXRs is slightly off.
-Because Blender does not expose the image resolution of multi layer exrs, we have to write out a jpg first and then check the resolution
-Currently only renders out single images not sequences