Add support to start media-viewer from cli with path list of media to open.
This commit reverts some changes from last commit and puts back most of the logic that was splitted out in the load_media operator back in to the callback_filename_change function as it is just more clear this way. The init_with_media_paths operator was added to process when the blender-media-viewer is started via the commandline and user supplies list of filepaths after the `--` argument. This operator basically just jumps to the right directory and selects the file defined by active_file_idx.

In the main __init__.py we parse the command line arguments and register the init_with_mediapaths load post handler when necessary.