media-viewer: add create MOVIE option for render review (image editor) op.
This patch adds an option to the render review image editor operator to create a movie file or image sequence. For realizing movie generation a second operator (convert_image_seq_to_movie) was added that starts a new blender instance to convert the image sequence to a movie file in the background.