Add vis module. Ensure obj visibility during data transfer.
Apparently Blender does not evaluate objects or collections in the depsgraph in some cases if they are not visible. This is something Users should not have to take care about when writing their transfer data instructions. So we will make sure here that everything is visible and after the transfer the original state will be restored.