Add UndoContext.
This is another Context object which you can add undo steps to. The point of it is to mainly be able to revert the filesystem and other things that happen between starting the publish and aborting it. These steps will also be mirrored on the scene poperty group so you can actualy start a publish open another scene and still abort it and it will undo the correct things.