SVN: Fix initializing in an untracked sub-folder.
The `svn info` command throws an error when called on a sub-folder of a repository
which is not yet tracked. However, this error is distinct from the error
we get when calling `svn info` outside of a repository, so we can simply
go up the folder hierarchy recursively until the command stops failing.
(Being mindful of avoiding infinite recursion in case the file
system root is an svn repository ofc.)