Add documentation system
This is the start of the Blender Studio Pipeline documentation
project. The goal is to describe the pipeline and tools used to
create the Blender Open Movie projects in a way that could be fully
replicated by everyone.

The documentation will be published at the Blender Studio website and
can be built locally by following the instructions in docs/README.md.