SVN: Restructure add-on
Remove unused util functions.
Remove svn_error dead code.

We only store svn_url and svn_directory in the Scene, which we use to match to the
corresponding SVN Repository data in the UserPrefs.
Replace `is_in_repo` flag with checking if a repository exists.
Move most properties from props.py to repository.py, which is stored in UserPrefs.
This allows storing the data of multiple repositories at once, allowing us to jump between
repositories in a future update.